Translates pagination aria-labels
